split auxd get rid of 'externalize crypto' warning
