+++ lxl-web/src/lib/utils/search.server.ts
@@ -35,7 +35,8 @@ import { type LocaleCode as LangCode } from '$lib/i18n/locales';
 import type { MyLibrariesType } from '$lib/types/userSettings';
 import { LxlLens } from '$lib/types/display';
 import { Width } from '$lib/types/auxd';
-import { bestImage, bestSize, toSecure } from '$lib/utils/auxd';
+import { bestImage, toSecure } from '$lib/utils/auxd.server';
+import { bestSize } from '$lib/utils/auxd';
 import getAtPath from '$lib/utils/getAtPath';
 import { getUriSlug } from '$lib/utils/http';
 import { isLibraryOrg } from '$lib/utils/holdings';
